requirements.txt
asgiref==3.5.2
certifi==2022.9.24
chainmap==1.0.3
charset-normalizer==2.1.1
combomethod==1.0.12
Django==4.1.3
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-gis==0.18
idna==3.4
nulltype==2.3.1
numpy==1.23.4
psycopg2-binary==2.9.5
python-decouple==3.6
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
django-crispy-forms==